// src/types/PhotoProps.ts
export interface PhotoProps {
    idFoto: number;
    title: string;
    file: string;
    description: string;
    number: number;
    lat: number;
    lon: number;
    seoDescription: string;
    location: string;
    nextPhoto: number;
    prevPhoto: number;
    nextTitle: string;
    prevTitle: string;
    sizeMaxW: number;
    sizeMaxH: number;
};